+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Fix wrong `check-coupled-files` behaviour when none of files from fileSets were changed
 
 ## [1.1.0] - 2017-12-04
 ### Added